Prevent scroll on initial load of the checkout
When loading the checkout the scrollBehavior was taking the User part way down the page so they didn't see the header and logo.

We can prevent this by checking the 'from' parameter and if it has a name. Without a name it means it came from a full page load so in this case take the User to the top of the page.