[HOSM-618] Fix call to undefined `getPaymentMethod` in 3DS failure
When closing or failing the 3DS check the payment method is cleared from the checkout but then we are trying to set the visible method back to that now cleared method which fails. Before calling the function `setToCurrentViewId()` we actually call a clear selected payment method which resets the view back to default so this call wasn't even needed.

I've also hardened the checks to ensure that Google Pay, PayPal and Venmo are available before changing their respective tokenize functions.