[HOSM-625] Allow validation of multiple reCAPTCHA types There's a quirk of Braintree that injects it's reCAPTCHA validation into all payment methods rather than just the Braintree methods.
What this change does is allow us to call `validateToken` with an array of placement IDs that are can be used for this placement. The method then gets the first enabled configuration and generates the token for that placement.

For example you can now call:

```
validateToken(['placeOrder', 'braintree'])
```

This will try to validate the token for placeOrder first and if that's not enabled we do it for Braintree afterwards.