Fixing an issue when getting the masked ID while not on the default store If you were on a non default store view as a logged in User, after proceeding to the checkout you would be immediately taken back to the cart page. This was because the graphQL request to get the masked ID would fail and the error handling on that redirects back to the cart page.
The reason for the failure was because the config hadn't been loaded yet so the default path for the grahpQL request would be `https://someurl.test/graphql` rather than the correct `https://someurl.test/[STORE_CODE]/graphql`.

Added another step within the CartStore so that we get the initial configuration before getting the masked ID. This way the correct graphQL URL is always known and used so that the request doesn't fail.

This has no impact on performance as the initial config request is cached locally on the frontend so it is only ever called once. All this change does is bring that request forward if it is needed.