Fix issues with getting domain with store code in URL When the store code is in the domain such as `https://mcstaging.skatehut.co.uk/eu` all REST requests should begin with this. However our helper was getting rid of that store code and using the very base path.
This means that REST requests would have been `https://mcstaging.skatehut.co.uk/rest/eu/V1/...` when the correct is `https://mcstaging.skatehut.co.uk/eu/rest/eu/V1/...`.

Updated the helper getting the base REST URL to use the store config value that Adobe Commerce provides as this contains the config `secureBaseLinkUrl` which is always the correct thing to use.

Also updated EmailAddress component as that was using the incorrect base URL too and always redirecting to the main website even if you were on a different store.