[HOSM-699] Fix issue with PayPal payment being creating with a 0.01 amount When we switched over to use the PayPal SDK the callback for createOrder wasn't being set so it defaults to the fallback version which sends a single penny value to PayPal.
Updated the patch to use a proper create order with the values that are passed in through configuration when creating the Braintree drop in.

Also the `onAuthorize` callback isn't used by PayPal's SDK and instead it uses the `onApprove` callback which we've duplicated so that the correct method is used.